requirements.txt
asn1crypto==0.24.0
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudinary==1.17.0
colorama==0.4.1
cryptography==2.7
Django==2.2.5
djangorestframework==3.10.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
PyJWT==1.7.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.2.0
oauthlib==3.0.1
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pylint==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.6.0
pytz==2019.2
requests==2.21.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.16.0
whitenoise==4.1.2
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
